Changed scheduler log to only include station positions
This means the log output reflects how long a train is at a station.  Before, the output was the time at which a train is told to move to a given location.  Time at station was not obvious with that implementation.  Also added the python file to plot the trajectories.